requirements.txt
streamlit
langgraph
langchain-google-genai
python-dotenv
PyPDF2
pdfplumber